backend/requirements.txt
Flask==3.1.0
python-dotenv
cryptography
IPFS-Toolkit
flask-cors
pyjwt
pymongo==4.11.3
apscheduler
atexit
python-dotenv